Rename toolbox to make it Ceph specific, short names for CRDs, upgrade guide tweaks
