+++ Documentation/toolbox.md
@@ -11,21 +11,21 @@ The toolbox is based on Ubuntu, so more tools of your choosing can be easily ins
 ## Running the Toolbox in Kubernetes
 
 The rook toolbox can run as a pod in a Kubernetes cluster.  After you ensure you have a running Kubernetes cluster with rook deployed (see the [Kubernetes](quickstart.md) instructions),
-launch the rook-tools pod.
+launch the rook-ceph-tools pod.
 
-Save the tools spec as `rook-tools.yaml`:
+Save the tools spec as `toolbox.yaml`:
 
 ```yaml
 apiVersion: v1
 kind: Pod
 metadata:
-  name: rook-tools
+  name: rook-ceph-tools
   namespace: rook-ceph
 spec:
   dnsPolicy: ClusterFirstWithHostNet
   containers:
-  - name: rook-tools
-    image: rook/toolbox:master
+  - name: rook-ceph-tools
+    image: rook/ceph-toolbox:master
     imagePullPolicy: IfNotPresent
     env:
       - name: ROOK_ADMIN_SECRET
@@ -63,19 +63,19 @@ spec:
           path: mon-endpoints
 ```
 
-Launch the rook-tools pod:
+Launch the rook-ceph-tools pod:
 ```bash
-kubectl create -f rook-tools.yaml
+kubectl create -f toolbox.yaml
 ```
 
 Wait for the toolbox pod to download its container and get to the `running` state:
 ```bash
-kubectl -n rook-ceph get pod rook-tools
+kubectl -n rook-ceph get pod rook-ceph-tools
 ```
 
-Once the rook-tools pod is running, you can connect to it with:
+Once the rook-ceph-tools pod is running, you can connect to it with:
 ```bash
-kubectl -n rook-ceph exec -it rook-tools bash
+kubectl -n rook-ceph exec -it rook-ceph-tools bash
 ```
 
 All available tools in the toolbox are ready for your troubleshooting needs.  Example:
@@ -88,7 +88,7 @@ rados df
 
 When you are done with the toolbox, remove the pod:
 ```bash
-kubectl -n rook-ceph delete pod rook-tools
+kubectl -n rook-ceph delete pod rook-ceph-tools
 ```
 
 ## Troubleshooting without the Toolbox
