+++ cluster/examples/kubernetes/ceph/operator.yaml
@@ -14,6 +14,8 @@ spec:
     listKind: ClusterList
     plural: clusters
     singular: cluster
+    shortNames:
+    - rcc
   scope: Namespaced
   version: v1alpha1
 ---
@@ -28,6 +30,8 @@ spec:
     listKind: FilesystemList
     plural: filesystems
     singular: filesystem
+    shortNames:
+    - rcfs
   scope: Namespaced
   version: v1alpha1
 ---
@@ -42,6 +46,8 @@ spec:
     listKind: ObjectStoreList
     plural: objectstores
     singular: objectstore
+    shortNames:
+    - rco
   scope: Namespaced
   version: v1alpha1
 ---
@@ -56,6 +62,8 @@ spec:
     listKind: PoolList
     plural: pools
     singular: pool
+    shortNames:
+    - rcp
   scope: Namespaced
   version: v1alpha1
 ---
@@ -70,6 +78,8 @@ spec:
     listKind: VolumeList
     plural: volumes
     singular: volume
+    shortNames:
+    - rv
   scope: Namespaced
   version: v1alpha2
 ---
