+++ cluster/examples/kubernetes/ceph/toolbox.yaml
@@ -1,13 +1,13 @@
 apiVersion: v1
 kind: Pod
 metadata:
-  name: rook-tools
+  name: rook-ceph-tools
   namespace: rook-ceph
 spec:
   dnsPolicy: ClusterFirstWithHostNet
   containers:
-  - name: rook-tools
-    image: rook/toolbox:master
+  - name: rook-ceph-tools
+    image: rook/ceph-toolbox:master
     imagePullPolicy: IfNotPresent
     env:
       - name: ROOK_ADMIN_SECRET
