+++ cmd/rook/ceph/ceph.go
@@ -58,6 +58,7 @@ func init() {
 
 func AddCommands(command *cobra.Command) {
 	command.AddCommand(operatorCmd)
+	command.AddCommand(agentCmd)
 	command.AddCommand(monCmd)
 	command.AddCommand(osdCmd)
 	command.AddCommand(mgrCmd)
