+++ images/ceph-toolbox/Dockerfile
