+++ images/ceph-toolbox/Dockerfile.base
