+++ images/ceph-toolbox/toolbox.sh
