+++ tests/scripts/makeTestImages.sh
@@ -9,13 +9,13 @@ case "${1:-}" in
     case "${2:-}" in
         arm|arm64|amd64)
             docker tag ${BUILD_REGISTRY}/ceph-$2:latest rook/ceph:master
-            docker tag ${BUILD_REGISTRY}/toolbox-$2:latest rook/toolbox:master
+            docker tag ${BUILD_REGISTRY}/ceph-toolbox-$2:latest rook/ceph-toolbox:master
             if [ ! -z $3 ]
                 then
                     docker tag ${BUILD_REGISTRY}/ceph-$2:latest rook/ceph:$3
-                    docker save -o ceph-$2.tar rook/ceph:master rook/ceph:$3 rook/toolbox:master
+                    docker save -o ceph-$2.tar rook/ceph:master rook/ceph:$3 rook/ceph-toolbox:master
                 else
-                   docker save -o ceph-$2.tar rook/ceph:master rook/toolbox:master
+                   docker save -o ceph-$2.tar rook/ceph:master rook/ceph-toolbox:master
             fi
             ;;
         *)
@@ -39,7 +39,7 @@ case "${1:-}" in
         arm|arm64|amd64)
             tag_version="${3:-"master"}"
             docker tag ${BUILD_REGISTRY}/ceph-$2:latest rook/ceph:${tag_version}
-            docker tag ${BUILD_REGISTRY}/toolbox-$2:latest rook/toolbox:${tag_version}
+            docker tag ${BUILD_REGISTRY}/ceph-toolbox-$2:latest rook/ceph-toolbox:${tag_version}
             ;;
         *)
             echo "usage :" >&2
